+++ packages/solid/src/methods/remove.ts
@@ -7,6 +7,7 @@ import {
   getFieldState,
   setFieldState,
   updateFieldArrayDirty,
+  validateIfNecessary,
 } from '../utils';
 
 type RemoveOptions = {
@@ -97,6 +98,9 @@ export function remove<
 
         // Update dirty state at field array and form
         updateFieldArrayDirty(form, fieldArray);
+
+        // Validate field array items if necessary
+        validateIfNecessary(form, name, { on: 'input' });
       }
     });
   });