+++ packages/solid/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Add `validateIfNecessary` util and refactor code
+- Fix missing validation at `insert` and `remove` method
 
 ## v0.6.0 (November 06, 2022)