+++ packages/solid/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the library will be documented in this file.
 
-## vX.X.X (Month DD, YYYY)
+## v0.7.0 (November 08, 2022)
 
 - Add `validateIfNecessary` util and refactor code
 - Fix missing validation at `insert` and `remove` method